requirements.txt
joblib==1.3.2
numpy==1.26.4
Pillow==9.0.1
h5py==3.11.0
scikit-learn==1.2.2
tensorflow==2.16.1
keras==3.3.3
streamlit==1.31.0